+++ mockserver-core/src/test/java/org/mockserver/matchers/HttpRequestsPropertiesMatcherTest.java
@@ -6095,7 +6095,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6117,7 +6117,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets/12345")
+            .withPath("/v1/pets/12345")
             .withHeader("X-Request-ID", UUID.randomUUID().toString())
             .withHeader("Other-Header", UUID.randomUUID().toString());
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
@@ -6140,7 +6140,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6158,6 +6158,72 @@ public class HttpRequestsPropertiesMatcherTest {
         httpRequestsPropertiesMatcher.update(new Expectation(
             new OpenAPIDefinition()
                 .withSpecUrlOrPayload("org/mockserver/openapi/openapi_petstore_example.yaml")
+                .withOperationId("listPets")
+        ));
+        HttpRequest httpRequest = request()
+            .withMethod("GET")
+            .withPath("/v1/pets")
+            .withQueryStringParameter("limit", "10");
+        MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
+
+        // when
+        boolean matches = httpRequestsPropertiesMatcher.matches(context, httpRequest);
+
+        // then
+        thenMatchesEmptyFieldDifferences(context, matches, true);
+    }
+
+    @Test
+    public void shouldMatchSingleOperationInOpenAPIWithOperationServerYamlUrl() {
+        // given
+        HttpRequestsPropertiesMatcher httpRequestsPropertiesMatcher = new HttpRequestsPropertiesMatcher(configuration, mockServerLogger);
+        httpRequestsPropertiesMatcher.update(new Expectation(
+            new OpenAPIDefinition()
+                .withSpecUrlOrPayload("org/mockserver/openapi/openapi_petstore_example_with_operation_server.yaml")
+                .withOperationId("listPets")
+        ));
+        HttpRequest httpRequest = request()
+            .withMethod("GET")
+            .withPath("/v2/pets")
+            .withQueryStringParameter("limit", "10");
+        MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
+
+        // when
+        boolean matches = httpRequestsPropertiesMatcher.matches(context, httpRequest);
+
+        // then
+        thenMatchesEmptyFieldDifferences(context, matches, true);
+    }
+
+    @Test
+    public void shouldMatchSingleOperationInOpenAPIWithServerContainingVariablesYamlUrl() {
+        // given
+        HttpRequestsPropertiesMatcher httpRequestsPropertiesMatcher = new HttpRequestsPropertiesMatcher(configuration, mockServerLogger);
+        httpRequestsPropertiesMatcher.update(new Expectation(
+            new OpenAPIDefinition()
+                .withSpecUrlOrPayload("org/mockserver/openapi/openapi_petstore_example_with_server_url_variables.yaml")
+                .withOperationId("listPets")
+        ));
+        HttpRequest httpRequest = request()
+            .withMethod("GET")
+            .withPath("/v2/pets")
+            .withQueryStringParameter("limit", "10");
+        MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
+
+        // when
+        boolean matches = httpRequestsPropertiesMatcher.matches(context, httpRequest);
+
+        // then
+        thenMatchesEmptyFieldDifferences(context, matches, true);
+    }
+
+    @Test
+    public void shouldMatchSingleOperationInOpenAPIWithServerHavingNoPathPrefixYamlUrl() {
+        // given
+        HttpRequestsPropertiesMatcher httpRequestsPropertiesMatcher = new HttpRequestsPropertiesMatcher(configuration, mockServerLogger);
+        httpRequestsPropertiesMatcher.update(new Expectation(
+            new OpenAPIDefinition()
+                .withSpecUrlOrPayload("org/mockserver/openapi/openapi_petstore_example_with_no_path_prefix.yaml")
                 .withOperationId("listPets")
         ));
         HttpRequest httpRequest = request()
@@ -6184,7 +6250,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6206,7 +6272,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6228,7 +6294,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6241,12 +6307,12 @@ public class HttpRequestsPropertiesMatcherTest {
         assertThat(context.getDifferences(PATH), containsInAnyOrder(
             "  string or regex match failed expected:" + NEW_LINE +
                 NEW_LINE +
-                "    /pets/.*" + NEW_LINE +
+                "    /v1/pets/.*" + NEW_LINE +
                 NEW_LINE +
                 "   found:" + NEW_LINE +
                 NEW_LINE +
-                "    /pets" + NEW_LINE,
-            "  expected path /pets/{petId} has 2 parts but path /pets has 1 part "
+                "    /v1/pets" + NEW_LINE,
+            "  expected path /v1/pets/{petId} has 3 parts but path /v1/pets has 2 parts "
         ));
         assertThat(context.getDifferences(QUERY_PARAMETERS), nullValue());
         assertThat(context.getDifferences(COOKIES), nullValue());
@@ -6269,7 +6335,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6291,7 +6357,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("PUT")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6342,12 +6408,12 @@ public class HttpRequestsPropertiesMatcherTest {
         assertThat(context.getDifferences(PATH), containsInAnyOrder(
             "  string or regex match failed expected:" + NEW_LINE +
                 NEW_LINE +
-                "    /pets/.*" + NEW_LINE +
+                "    /v1/pets/.*" + NEW_LINE +
                 NEW_LINE +
                 "   found:" + NEW_LINE +
                 NEW_LINE +
                 "    /wrong" + NEW_LINE,
-            "  expected path /pets/{petId} has 2 parts but path /wrong has 1 part "
+            "  expected path /v1/pets/{petId} has 3 parts but path /wrong has 1 part "
         ));
         assertThat(context.getDifferences(QUERY_PARAMETERS), nullValue());
         assertThat(context.getDifferences(COOKIES), nullValue());
@@ -6370,7 +6436,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/some/path")
+            .withPath("/v1/some/path")
             .withQueryStringParameter("limit", "10");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6423,7 +6489,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/some/path")
+            .withPath("/v1/some/path")
             .withQueryStringParameter("limit", "not_a_number");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6475,7 +6541,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("GET")
-            .withPath("/some/path")
+            .withPath("/v1/some/path")
             .withQueryStringParameter("limit", "not_a_number");
         MatchDifference context = new MatchDifference(configuration.detailedMatchFailures(), httpRequest);
 
@@ -6486,18 +6552,18 @@ public class HttpRequestsPropertiesMatcherTest {
         assertThat(matches, is(false));
         assertThat(context.getDifferences(PATH), containsInAnyOrder("  string or regex match failed expected:" + NEW_LINE +
                 NEW_LINE +
-                "    /pets" + NEW_LINE +
+                "    /v1/pets" + NEW_LINE +
                 NEW_LINE +
                 "   found:" + NEW_LINE +
                 NEW_LINE +
-                "    /some/path" + NEW_LINE,
+                "    /v1/some/path" + NEW_LINE,
             "  string or regex match failed expected:" + NEW_LINE +
                 NEW_LINE +
-                "    /pets/.*" + NEW_LINE +
+                "    /v1/pets/.*" + NEW_LINE +
                 NEW_LINE +
                 "   found:" + NEW_LINE +
                 NEW_LINE +
-                "    /some/path" + NEW_LINE));
+                "    /v1/some/path" + NEW_LINE));
         assertThat(context.getDifferences(METHOD), containsInAnyOrder("  string or regex match failed expected:" + NEW_LINE +
                 NEW_LINE +
                 "    POST" + NEW_LINE +
@@ -6554,7 +6620,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("POST")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withHeader("content-type", "application/json")
             .withBody(json("{" + NEW_LINE +
                 "    \"id\": \"invalid_id_format\", " + NEW_LINE +
@@ -6622,7 +6688,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("POST")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withHeader("content-type", "application/json")
             .withBody(json("{" + NEW_LINE +
                 "    \"id\": \"invalid_id_format\", " + NEW_LINE +
@@ -6698,7 +6764,7 @@ public class HttpRequestsPropertiesMatcherTest {
         ));
         HttpRequest httpRequest = request()
             .withMethod("POST")
-            .withPath("/pets")
+            .withPath("/v1/pets")
             .withBody(json("{" + NEW_LINE +
                 "    \"id\": \"invalid_id_format\", " + NEW_LINE +
                 "    \"name\": \"scruffles\", " + NEW_LINE +