+++ mockserver-integration-testing/src/main/java/org/mockserver/testing/integration/mock/AbstractBasicMockingIntegrationTest.java
@@ -585,7 +585,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath("/pets")
+                    .withPath("/v1/pets")
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE)
         );
@@ -622,7 +622,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath("/pets")
+                    .withPath("/v1/pets")
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE)
         );
@@ -638,7 +638,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("POST")
-                    .withPath("/pets")
+                    .withPath("/v1/pets")
                     .withBody(json("{" + NEW_LINE +
                         "  \"id\" : 0," + NEW_LINE +
                         "  \"name\" : \"some_string_value\"," + NEW_LINE +
@@ -659,7 +659,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath("/pets/12345")
+                    .withPath("/v1/pets/12345")
                     .withHeader("x-request-id", UUIDService.getUUID()),
                 HEADERS_TO_IGNORE)
         );
@@ -949,7 +949,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("PUT")
-                    .withPath("/pets")
+                    .withPath("/v1/pets")
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE)
         );
@@ -1281,7 +1281,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE
             )
@@ -1291,7 +1291,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("POST")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withHeader("content-type", "application/json")
                     .withBody(json("" +
                         "{" + NEW_LINE +
@@ -1493,7 +1493,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE
             )
@@ -1503,7 +1503,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("POST")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withHeader("content-type", "application/json")
                     .withBody(json("" +
                         "{" + NEW_LINE +
@@ -1726,7 +1726,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE
             )
@@ -1743,7 +1743,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("POST")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withHeader("content-type", "application/json")
                     .withBody(json("" +
                         "{" + NEW_LINE +
@@ -1762,11 +1762,11 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             mockServerClient.retrieveRecordedRequests(openAPI().withSpecUrlOrPayload(FileReader.readFileFromClassPathOrPath("org/mockserver/openapi/openapi_petstore_example.json"))),
             request()
                 .withMethod("GET")
-                .withPath("/pets")
+                .withPath("/v1/pets")
                 .withQueryStringParameter("limit", "10"),
             request()
                 .withMethod("POST")
-                .withPath("/pets")
+                .withPath("/v1/pets")
                 .withHeader("content-type", "application/json")
                 .withBody(json("" +
                     "{" + NEW_LINE +
@@ -2259,7 +2259,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE
             )
@@ -2276,7 +2276,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("POST")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withHeader("content-type", "application/json")
                     .withBody(json("" +
                         "{" + NEW_LINE +
@@ -2320,7 +2320,7 @@ public abstract class AbstractBasicMockingIntegrationTest extends AbstractMockin
             makeRequest(
                 request()
                     .withMethod("GET")
-                    .withPath(calculatePath("/pets"))
+                    .withPath(calculatePath("/v1/pets"))
                     .withQueryStringParameter("limit", "10"),
                 HEADERS_TO_IGNORE
             )